Spinner legend
Abstracted spinner legend logic away from spinner display logic. I plan to genericize the spinner display logic.
